+++ packages/core/build-vendor.sh
@@ -7,3 +7,4 @@ cp -r ../../node_modules/@carbon/motion src/styles/vendor/@carbon
 cp -r ../../node_modules/@carbon/type src/styles/vendor/@carbon
 cp -r ../../node_modules/@carbon/themes src/styles/vendor/@carbon
 cp -r ../../node_modules/@carbon/colors src/styles/vendor/@carbon
+cp -r ../../node_modules/carbon-components src/styles/vendor