+++ packages/core/src/components/essentials/tooltip-bar.ts
@@ -26,19 +26,21 @@ export class TooltipBar extends Tooltip {
 			if ((e.detail.type === TooltipTypes.DATAPOINT && Tools.getProperty(this.model.getOptions(), "tooltip", "datapoint", "enabled"))
 				|| (e.detail.type === TooltipTypes.GRIDLINE && Tools.getProperty(this.model.getOptions(), "tooltip", "gridline", "enabled")) ) {
 
+				let data = e.detail.hoveredElement.datum() as any;
 				const hoveredElement = e.detail.hoveredElement.node();
 
 				let defaultHTML;
 				if (e.detail.multidata) {
 					// multi tooltip
-					defaultHTML = this.getMultilineTooltipHTML(e.detail.multidata);
+					data = e.detail.multidata;
+					defaultHTML = this.getMultilineTooltipHTML(data);
 				} else {
 					defaultHTML = this.getTooltipHTML(e.detail.hoveredElement.datum());
 				}
 
 				// if there is a provided tooltip HTML function call it and pass the defaultHTML
 				if (Tools.getProperty(this.model.getOptions(), "tooltip", "customHTML")) {
-					tooltipTextContainer.html(this.model.getOptions().tooltip.customHTML(hoveredElement, defaultHTML));
+					tooltipTextContainer.html(this.model.getOptions().tooltip.customHTML(data, defaultHTML));
 				} else {
 					// default tooltip
 					tooltipTextContainer.html(defaultHTML);
@@ -142,7 +144,8 @@ export class TooltipBar extends Tooltip {
 				const formattedValue = Tools.getProperty(this.model.getOptions(), "tooltip", "valueFormatter") ?
 				this.model.getOptions().tooltip.valueFormatter(datapoint.value) : datapoint.value.toLocaleString("en");
 
-				const indicatorColor = this.model.getStrokeColor(datapoint.datasetLabel, datapoint.label, datapoint.value);
+				// For the tooltip color, we always want the normal stroke color, not dynamically determined by data value.
+				const indicatorColor = this.model.getStrokeColor(datapoint.datasetLabel, datapoint.label);
 
 				return `
 				<li>