+++ packages/core/src/interfaces/charts.ts
@@ -30,13 +30,17 @@ export interface BaseChartOptions {
 	 */
 	legend?: LegendOptions;
 	/**
-	 * Optional function to generate the fill color based on datasetLabel, label, and/or value
+	 * Optional function to determine whether is filled based on datasetLabel, label, and/or data
 	 */
-	getFillColor?: (datasetLabel: any, label?: any, value?: any) => string;
+	getIsFilled?: (datasetLabel: any, label?: any, data?: any, defaultFilled?: boolean) => boolean;
 	/**
-	 * Optional function to generate the stroke color based on datasetLabel, label, and/or value
+	 * Optional function to generate the fill color based on datasetLabel, label, and/or data
 	 */
-	getStrokeColor?: (datasetLabel: any, label?: any, value?: any) => string;
+	getFillColor?: (datasetLabel: any, label?: any, data?: any, defaultFillColor?: string) => string;
+	/**
+	 * Optional function to generate the stroke color based on datasetLabel, label, and/or data
+	 */
+	getStrokeColor?: (datasetLabel: any, label?: any, data?: any, defaultStrokeColor?: string) => string;
 	/**
 	 * stylesheet options
 	 */